Project_2/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
python-dateutil==2.8.1
pytz==2021.1
sqlparse==0.4.1
